fix: use 32 bit int and float in `MAAValue` (#198)
If we pass a big int (larger than i32::MAX) to core, it will crash.

+++ maa-cli/src/main.rs
@@ -170,7 +170,7 @@ enum SubCommand {
         stage: String,
         /// medicine to use
         #[arg(short, long)]
-        medicine: Option<i64>,
+        medicine: Option<i32>,
         #[command(flatten)]
         common: run::CommonArgs,
     },
